+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 ### Added (picodata)
 - struct `read_view::ReadView` for opening read views on selected spaces.
 - Expose `box_access_check_space` to be able to run access checks on spaces externally
+- ffi logger functions `log_set_format` and `log_default_logger`
 
 ### Fixed (picodata)
 - A race condition causing undefined behaviour due to fiber_cond_delete being called outside tx sometimes